pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.3"
gunicorn = "^22.0.0"
uvicorn-worker = "^0.2.0"
prometheus-fastapi-instrumentator = "^7.0.0"
sentry-sdk = "^2.5.1"
fastapi-utils = "^0.7.0"
typing-inspect = "^0.9.0"
sqlalchemy = "^2.0.31"
psycopg = {extras = ["binary"], version = "^3.1.19"}
factory-boy = "^3.3.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
loguru = "^0.7.2"
slowapi = "^0.1.9"
redis = "^5.0.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
mypy = "^1.10.0"
ruff = "^0.4.9"
bandit = "^1.7.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.ruff]
target-version = "py312"
line-length = 99
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
indent-width = 4

[tool.ruff.lint]
# Select which codes to check
# E: Pyflakes (error codes)
# F: PyFlakes (additional checks)
# W: PEP8 (warning codes)
# B: Bugbear (additional checks for potential issues)
# A: flake8-builtins
# SIM: flake8-simplify
# DTZ: flake8-datetimez
select = ["E", "F", "W", "B", "A", "SIM", "DTZ"]
ignore = [
    "SIM105",
    "SIM116",
]
exclude = [
    "app/tests/*",  # Test files
    "venv",        # Virtual environments
    ".venv",       # Virtual environments
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Disallow unused variables when underscore-prefixed.
dummy-variable-rgx = ""

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E401", "F401"]



[tool.black]
line-length = 99 # same as Ruff
target-version = ['py312']
skip-string-normalization = false
skip-magic-trailing-comma = false
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | \venv
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
pretty = true
allow_untyped_decorators = true
exclude = ["app/tests"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
version_files = [ "pyproject.toml:version"]

[tool.commitizen.customize]
schema_pattern  = "(?s)^(feat|fix|enhancement|ref|style|test|docs|build|ops|chore|revert|bumps)(\\(\\S+\\))?!?: [^\\n\\r]+((\\n\\n.*)|(\\s*))?$"

[tool.bandit]
targets = ["app/"]
exclude = ["app/tests/"]
profile = "medium"
options = ["--recursive"]
skips = [
    "B201",
    "B610",
    "B611",
    "B703"
]